Extract utility functions to submodule
To improve readability of the Skill and separate concerns.
